veracruz-utils-reorganization: reorganization of content of veracruz-utils
This is to make way for Linux support.  Veracruz utils has become a bit of an unruly mess, with lots of material that is not well organized, and some large monolithic files containing all sorts of poorly-related material.

    1. Move platform-specific material under a dedicated `platform` module hierarchy, with submodules for each platform (e.g. `tz`) and in-time related groups of platforms (e.g. `vm` for VM-related platforms like Nitro and Linux).
    2. Split some large monlothic files up in `policy.rs` into smaller, more focussed source files containing closely-related material.
    3. Introduce a new module, `io`, dedicated to material related-to vsockets and rawfd types.  Some material has been moved out of Nitro-related files as a result.
    4. Made some small generalizations and quality-of-life improvements to existing material, e.g. policy timepoints are now checked as they are constructed to make sure that month, days, hours, minutes look reasonable.  Previously a TODO.  Preferred use of Rust stdlib `TryFrom` crate instead of ad-hoc conversion functions from `u32` in various opcode files, and so on.
    5. Added new Rustdoc comments throughout.
    6. Stopped re-exporting all modules from `veracruz-utils`.
