+++ veracruz-server/fuzz/fuzz_targets/init_veracruz_server.rs
@@ -12,7 +12,7 @@
 #![no_main]
 use libfuzzer_sys::fuzz_target;
 
-use veracruz_utils;
+use veracruz_utils::policy::policy::Policy;
 // use SGX to fuzz the functionality
 use veracruz_server::VeracruzServerSGX as VeracruzServerEnclave;
 use veracruz_server::veracruz_server::VeracruzServer;
@@ -20,7 +20,7 @@ use veracruz_server::veracruz_server::VeracruzServer;
 fuzz_target!(|data: &[u8]| {
     // fuzzed code goes here
     if let Ok(s) = std::str::from_utf8(data) {
-        if let Ok(policy) = veracruz_utils::VeracruzPolicy::new(s) {
+        if let Ok(policy) = Policy::new(s) {
             let enclave = VeracruzServerEnclave::new(&policy);
             assert!(enclave.is_ok());
         }
