+++ veracruz-utils/Cargo.toml
@@ -15,7 +15,7 @@ nitro = ["serde_json/std", "nix", "byteorder"]
 [dependencies]
 sgx_tstd = { rev = "v1.1.2", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
 serde = { git = "https://github.com/veracruz-project/serde.git", features=["derive"], branch = "veracruz" }
-x509-parser={ version = "0.7.0", optional = true }
+x509-parser = { version = "0.7.0", optional = true }
 serde_json = { git = "https://github.com/veracruz-project/json.git", branch = "veracruz", default-features = false }
 # The cargo patch mechanism does NOT work when we add function into a macro_rules!
 rustls = { git = "https://github.com/veracruz-project/rustls.git", branch = "veracruz", optional = true }
