+++ veracruz-utils/src/platform/nitro/mod.rs
@@ -1,0 +1,13 @@
+//! AWS Nitro enclaves platform-specific material.
+//!
+//! ## Authors
+//!
+//! The Veracruz Development Team.
+//!
+//! ## Licensing and copyright notice
+//!
+//! See the `LICENSE.markdown` file in the Veracruz root directory for
+//! information on licensing and copyright.
+
+pub mod nitro;
+pub mod nitro_enclave;
