+++ veracruz-utils/src/platform/nitro/nitro_enclave.rs
@@ -1,6 +1,6 @@
 //! Nitro-Enclave-specific material for Veracruz
 //!
-//! ## Authors
+//! ## Authors
 //!
 //! The Veracruz Development Team.
 //!
@@ -10,16 +10,16 @@
 //! information on licensing and copyright.
 
 use err_derive::Error;
-use nix::sys::socket::sockopt::{ReuseAddr, ReusePort};
-use nix::sys::socket::{
-    accept, bind, listen, setsockopt, shutdown, socket, AddressFamily, Shutdown, SockAddr,
-    SockFlag, SockType,
+use nix::{
+    sys::socket::{
+        accept, bind, listen, setsockopt, shutdown, socket,
+        sockopt::{ReuseAddr, ReusePort},
+        AddressFamily, Shutdown, SockAddr, SockFlag, SockType,
+    },
+    unistd::close
 };
-use nix::unistd::close;
 use serde_json::Value;
-use std::os::unix::io::AsRawFd;
-use std::process::Command;
-use std::thread::JoinHandle;
+use std::{os::unix::io::AsRawFd, process::Command, thread::JoinHandle};
 
 /// Errors generated by Nitro enclave components of Veracruz
 #[derive(Debug, Error)]
@@ -41,7 +41,7 @@ pub enum NitroError {
     CLIError,
     /// a Veracruz-specific socket error occurred
     #[error(display = "Nitro: Veracruz Socket Error:{:?}", _0)]
-    VeracruzSocketError(#[error(source)] crate::VeracruzSocketError),
+    VeracruzSocketError(#[error(source)] crate::io::error::SocketError),
     /// An error occured while processing UTF8 string data
     #[error(display = "Nitro: Utf8Error:{:?}", _0)]
     Utf8Error(#[error(source)] std::str::Utf8Error),
@@ -60,13 +60,13 @@ pub struct NitroEnclave {
     /// value
     enclave_id: String,
     /// A convenience struct for handling VSOCK connections to the enclave
-    vsocksocket: crate::vsocket::VsockSocket,
+    vsocksocket: crate::io::vsocket::VsockSocket,
     /// The thread that will handle ocall operations from the enclave, if enabled
     /// (hence it's an `Option`
     ocall_thread: Option<JoinHandle<()>>,
     /// A communication channel to tell the ocall thread to terminate (if enabled)
     ocall_terminate_sender: Option<std::sync::Mutex<std::sync::mpsc::Sender<bool>>>,
-    /// the path to the Nictro CLI function. Not all AMI images have it in the 
+    /// the path to the Nictro CLI function. Not all AMI images have it in the
     /// same place in the file system, so we need to keep track of it
     nitro_cli_path: String,
 }
@@ -77,7 +77,7 @@ const VERACRUZ_PORT: u32 = 5005;
 const VMADDR_CID_ANY: u32 = 0xFFFFFFFF;
 /// The port that the ocall thread will be waiting for socket connections on
 const OCALL_PORT: u32 = 5006;
-/// the maximum back log of events that will be maintained while listening to 
+/// the maximum back log of events that will be maintained while listening to
 /// a socket
 const BACKLOG: usize = 128;
 
@@ -168,7 +168,7 @@ impl NitroEnclave {
                 .to_string()
                 .trim_matches('"')
                 .to_string(),
-            vsocksocket: crate::vsocket::vsock_connect(cid, VERACRUZ_PORT)?,
+            vsocksocket: crate::io::vsocket::VsockSocket::connect(cid, VERACRUZ_PORT)?,
             ocall_thread: ocall_thread_opt,
             ocall_terminate_sender: sender,
             nitro_cli_path: nitro_cli_path.to_string(),
@@ -236,12 +236,12 @@ impl NitroEnclave {
             loop {
                 match accept(socket_fd) {
                     Ok(fd) => {
-                        let received_buffer = crate::nitro::receive_buffer(fd)
+                        let received_buffer = crate::io::raw_fd::receive_buffer(fd)
                             .expect("NitroEnclave::ocall_loop failed to receive buffer");
                         // call the handler
                         let return_buffer = handler(received_buffer)
                             .expect("NitroEnclave::ocall_loop handler failed");
-                        crate::nitro::send_buffer(fd, &return_buffer)
+                        crate::io::raw_fd::send_buffer(fd, &return_buffer)
                             .expect("NitroEnclave::ocall_loop failed to send buffer");
                     }
                     Err(err) => match err {
@@ -264,10 +264,10 @@ impl NitroEnclave {
                 nix::Error::Sys(err) => {
                     match err {
                         nix::errno::Errno::ENOTCONN => (), // this is normal
-                        _ =>  println!("NitroEnclave::ocall_loop failed to shutdown socket({:?}. This might cause you problems in the future.", err),
+                        _ =>  println!("NitroEnclave::ocall_loop failed to shutdown socket ({:?}. This might cause you problems in the future.", err),
                     }
                 },
-                _ => println!("NitroEnclave::ocall_loop failed to shutdown socket({:?}. This might cause you problems in the future.", err),
+                _ => println!("NitroEnclave::ocall_loop failed to shutdown socket ({:?}. This might cause you problems in the future.", err),
             }
         }
         if let Err(err) = close(socket_fd) {
@@ -277,13 +277,13 @@ impl NitroEnclave {
 
     /// send a buffer of data to the enclave
     pub fn send_buffer(&self, buffer: &Vec<u8>) -> Result<(), NitroError> {
-        crate::nitro::send_buffer(self.vsocksocket.as_raw_fd(), buffer)
+        crate::io::raw_fd::send_buffer(self.vsocksocket.as_raw_fd(), buffer)
             .map_err(|err| NitroError::VeracruzSocketError(err))
     }
 
     /// receive a buffer of data from the enclave
     pub fn receive_buffer(&self) -> Result<Vec<u8>, NitroError> {
-        crate::nitro::receive_buffer(self.vsocksocket.as_raw_fd())
+        crate::io::raw_fd::receive_buffer(self.vsocksocket.as_raw_fd())
             .map_err(|err| NitroError::VeracruzSocketError(err))
     }
 }
