+++ veracruz-utils/src/platform/tz/mod.rs
@@ -1,0 +1,13 @@
+//! Arm TrustZone platform-specific material.
+//!
+//! ## Authors
+//!
+//! The Veracruz Development Team.
+//!
+//! ## Licensing and copyright notice
+//!
+//! See the `LICENSE.markdown` file in the Veracruz root directory for
+//! information on licensing and copyright.
+
+pub mod root_enclave_opcode;
+pub mod runtime_manager_opcode;
