+++ veracruz-utils/src/policy/mod.rs
@@ -1,0 +1,35 @@
+//! Types and definitions relating to the Veracruz global policy.
+//!
+//! The global policy captures important information about a Veracruz
+//! computation that principals need to audit before they enroll themselves in a
+//! computation.  This includes:
+//!
+//! - The identities and roles of every principals in the computation,
+//! - Important URLs, both for the Veracruz bridge server on the untrusted
+//!   host's machine and the Veracruz proxy attestation service,
+//! - Permissible ciphersuites for TLS connections between clients and the
+//!   trusted Veracruz runtime, as well as the hashes of the expected program
+//!   and of the trusted Veracruz runtime itself,
+//! - The expiry date (moment in time) of the self-signed certificate issued by
+//!   the enclave during a pre-computation bootstrapping process,
+//! - The execution strategy that will be used by the trusted Veracruz runtime
+//!   to execute the WASM binary, as well as a debug configuration flag which
+//!   allows the WASM binary to write data to `stdout` on the untrusted host's
+//!   machine,
+//! - The order in which data inputs provisioned into the enclave will be placed
+//!   which is important for the program provider to understand in order to
+//!   write software for Veracruz.
+//!
+//! ## Authors
+//!
+//! The Veracruz Development Team.
+//!
+//! ## Licensing and copyright notice
+//!
+//! See the `LICENSE.markdown` file in the Veracruz root directory for
+//! information on licensing and copyright.
+
+pub mod error;
+pub mod expiry;
+pub mod policy;
+pub mod principal;
